Add ThreadLocalCache for improved memory management
Implemented a new `ThreadLocalCache` class to manage memory allocations for commands in a thread-local manner. This includes methods for allocating memory, resetting the cache, and freeing all allocated memory.

Updated `CommandAllocator` to integrate `ThreadLocalCache`, enhancing memory management efficiency. Refactored code to use modern C++ features for better thread safety and performance.

Restructured `CMakeLists.txt` to add and remove relevant files, promoting better organization of the codebase. Overall, these changes improve memory management capabilities in multi-threaded scenarios, reducing contention and enhancing allocation performance.
