+++ OloEngine/src/CMakeLists.txt
@@ -130,20 +130,22 @@
 		
 		"OloEngine/Renderer/RenderCommands/RenderCommandBase.h"
 		
-		"OloEngine/Renderer/Commands/RenderCommand.h"
-		"OloEngine/Renderer/Commands/CommandPacket.h"
-		"OloEngine/Renderer/Commands/CommandPacket.cpp"
-		"OloEngine/Renderer/Commands/CommandPacketList.h"
-		"OloEngine/Renderer/Commands/CommandPacketList.cpp"
-		"OloEngine/Renderer/Commands/CommandRingBuffer.h"
-		"OloEngine/Renderer/Commands/CommandRingBuffer.cpp"
-		"OloEngine/Renderer/Commands/CommandBuffer.h"
-		"OloEngine/Renderer/Commands/CommandBuffer.cpp"
-		"OloEngine/Renderer/Commands/CommandDispatch.h"
-		"OloEngine/Renderer/Commands/CommandAllocator.h"
 		"OloEngine/Renderer/Commands/CommandAllocator.cpp"
-		"OloEngine/Renderer/Commands/CommandMemoryManager.h"
+		"OloEngine/Renderer/Commands/CommandAllocator.h"
+		"OloEngine/Renderer/Commands/CommandBuffer.cpp"
+		"OloEngine/Renderer/Commands/CommandBuffer.h"
+		"OloEngine/Renderer/Commands/CommandDispatch.h"
 		"OloEngine/Renderer/Commands/CommandMemoryManager.cpp"
+		"OloEngine/Renderer/Commands/CommandMemoryManager.h"
+		"OloEngine/Renderer/Commands/CommandPacket.cpp"
+		"OloEngine/Renderer/Commands/CommandPacket.h"
+		"OloEngine/Renderer/Commands/CommandPacketList.cpp"
+		"OloEngine/Renderer/Commands/CommandPacketList.h"
+		"OloEngine/Renderer/Commands/CommandRingBuffer.cpp"
+		"OloEngine/Renderer/Commands/CommandRingBuffer.h"
+		"OloEngine/Renderer/Commands/RenderCommand.h"
+		"OloEngine/Renderer/Commands/ThreadLocalCache.cpp"
+		"OloEngine/Renderer/Commands/ThreadLocalCache.h"
 
 		"OloEngine/Scene/Components.h"
 		"OloEngine/Scene/Entity.cpp"
